+++ metadata/src/main/java/org/apache/kafka/image/TopicImage.java
@@ -35,7 +35,7 @@ import java.util.Map.Entry;
  *
  * This class is thread-safe.
  */
-public record TopicImage(String name, Uuid id, String mirrorName, int desiredMirrorState, Map<Integer, PartitionRegistration> partitions) {
+public record TopicImage(String name, Uuid id, String mirrorName, byte desiredMirrorState, Map<Integer, PartitionRegistration> partitions) {
     public TopicImage(String name, Uuid id, Map<Integer, PartitionRegistration> partitions) {
         this(name, id, null, MirrorPartitionState.UNKNOWN.value(), partitions);
     }
@@ -52,7 +52,7 @@ public record TopicImage(String name, Uuid id, String mirrorName, int desiredMir
             writer.write(0, new MirrorTopicStateChangeRecord().
                     setTopicId(id).
                     setMirrorName(mirrorName).
-                    setDesiredState((byte) desiredMirrorState));
+                    setDesiredState(desiredMirrorState));
         }
         for (Entry<Integer, PartitionRegistration> entry : partitions.entrySet()) {
             int partitionId = entry.getKey();
